Coerce to a tibble or data frame

Coerce to a tibble or data frame

# S3 method for tbl_ts
as_tibble(x, ...)

# S3 method for tbl_ts
as.data.frame(x, row.names = NULL, optional = FALSE,
  ...)

Arguments

x
A tbl_ts.

...
Ignored.

row.names
NULL or a character vector giving the row names for the data frame. Missing values are not allowed.

optional
logical. If TRUE, setting row names and converting column names (to syntactic names: see make.names) is optional. Note that all of R's base package as.data.frame() methods use optional only for column names treatment, basically with the meaning of data.frame(*, check.names = !optional). See also the make.names argument of the matrix method.

Examples

as_tibble(pedestrian)

#> [38;5;246m# A tibble: 66,037 x 5[39m
#>    Sensor         Date_Time           Date        Time Count
#>    [3m[38;5;246m<chr>[39m[23m          [3m[38;5;246m<dttm>[39m[23m              [3m[38;5;246m<date>[39m[23m     [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<int>[39m[23m
#> [38;5;250m 1[39m Birrarung Marr 2015-01-01 [38;5;246m00:00:00[39m 2015-01-01     0  [4m1[24m630
#> [38;5;250m 2[39m Birrarung Marr 2015-01-01 [38;5;246m01:00:00[39m 2015-01-01     1   826
#> [38;5;250m 3[39m Birrarung Marr 2015-01-01 [38;5;246m02:00:00[39m 2015-01-01     2   567
#> [38;5;250m 4[39m Birrarung Marr 2015-01-01 [38;5;246m03:00:00[39m 2015-01-01     3   264
#> [38;5;250m 5[39m Birrarung Marr 2015-01-01 [38;5;246m04:00:00[39m 2015-01-01     4   139
#> [38;5;250m 6[39m Birrarung Marr 2015-01-01 [38;5;246m05:00:00[39m 2015-01-01     5    77
#> [38;5;250m 7[39m Birrarung Marr 2015-01-01 [38;5;246m06:00:00[39m 2015-01-01     6    44
#> [38;5;250m 8[39m Birrarung Marr 2015-01-01 [38;5;246m07:00:00[39m 2015-01-01     7    56
#> [38;5;250m 9[39m Birrarung Marr 2015-01-01 [38;5;246m08:00:00[39m 2015-01-01     8   113
#> [38;5;250m10[39m Birrarung Marr 2015-01-01 [38;5;246m09:00:00[39m 2015-01-01     9   166
#> [38;5;246m# ... with 66,027 more rows[39m

# a grouped tbl_ts -----
grped_ped <- pedestrian %>% group_by(Sensor)
as_tibble(grped_ped)

#> [38;5;246m# A tibble: 66,037 x 5[39m
#> [38;5;246m# Groups:   Sensor [4][39m
#>    Sensor         Date_Time           Date        Time Count
#>    [3m[38;5;246m<chr>[39m[23m          [3m[38;5;246m<dttm>[39m[23m              [3m[38;5;246m<date>[39m[23m     [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<int>[39m[23m
#> [38;5;250m 1[39m Birrarung Marr 2015-01-01 [38;5;246m00:00:00[39m 2015-01-01     0  [4m1[24m630
#> [38;5;250m 2[39m Birrarung Marr 2015-01-01 [38;5;246m01:00:00[39m 2015-01-01     1   826
#> [38;5;250m 3[39m Birrarung Marr 2015-01-01 [38;5;246m02:00:00[39m 2015-01-01     2   567
#> [38;5;250m 4[39m Birrarung Marr 2015-01-01 [38;5;246m03:00:00[39m 2015-01-01     3   264
#> [38;5;250m 5[39m Birrarung Marr 2015-01-01 [38;5;246m04:00:00[39m 2015-01-01     4   139
#> [38;5;250m 6[39m Birrarung Marr 2015-01-01 [38;5;246m05:00:00[39m 2015-01-01     5    77
#> [38;5;250m 7[39m Birrarung Marr 2015-01-01 [38;5;246m06:00:00[39m 2015-01-01     6    44
#> [38;5;250m 8[39m Birrarung Marr 2015-01-01 [38;5;246m07:00:00[39m 2015-01-01     7    56
#> [38;5;250m 9[39m Birrarung Marr 2015-01-01 [38;5;246m08:00:00[39m 2015-01-01     8   113
#> [38;5;250m10[39m Birrarung Marr 2015-01-01 [38;5;246m09:00:00[39m 2015-01-01     9   166
#> [38;5;246m# ... with 66,027 more rows[39m