Coerce to a tsibble object

Coerce to a tsibble object

as_tsibble(x, ...)

# S3 method for tbl_df
as_tsibble(x, key = id(), index, regular = TRUE,
  validate = TRUE, ...)

# S3 method for tbl_ts
as_tsibble(x, validate = FALSE, ...)

# S3 method for data.frame
as_tsibble(x, key = id(), index, regular = TRUE,
  validate = TRUE, ...)

# S3 method for list
as_tsibble(x, key = id(), index, regular = TRUE,
  validate = TRUE, ...)

# S3 method for ts
as_tsibble(x, tz = "UTC", ...)

# S3 method for mts
as_tsibble(x, tz = "UTC", gather = TRUE, ...)

# S3 method for msts
as_tsibble(x, tz = "UTC", gather = TRUE, ...)

# S3 method for hts
as_tsibble(x, tz = "UTC", ...)

Arguments

x
Other objects to be coerced to a tsibble (tbl_ts).

...
Other arguments passed on to individual methods.

key
Variable(s) that define unique time indices, used with the helper id(). If a univariate time series (without an explicit key), simply call id().

index
A bare (or unquoted) variable to specify the time index variable.

regular
Regular time interval (TRUE) or irregular (FALSE). The interval is determined by the greatest common divisor of positive time distances, if TRUE.

validate
TRUE suggests to verify that each key or each combination of key variables lead to unique time indices (i.e. a valid tsibble). If you are sure that it's a valid input, specify FALSE to skip the checks.

tz
Time zone. May be useful when a ts object is more frequent than daily.

gather
TRUE gives a "long" data form, otherwise as "wide" as x.

Value

A tsibble object.

See also

Examples

# coerce tibble to tsibble w/o a key ----
tbl1 <- tibble(
  date = as.Date("2017-01-01") + 0:9,
  value = rnorm(10)
)
as_tsibble(tbl1)

#> Column `date` is the index.

#> [38;5;246m# A tsibble: 10 x 2 [1D][39m
#>    date        value
#>    [3m[38;5;246m<date>[39m[23m      [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m 1[39m 2017-01-01 -[31m1[39m[31m.[39m[31m82[39m 
#> [38;5;250m 2[39m 2017-01-02 -[31m0[39m[31m.[39m[31m247[39m
#> [38;5;250m 3[39m 2017-01-03 -[31m0[39m[31m.[39m[31m244[39m
#> [38;5;250m 4[39m 2017-01-04 -[31m0[39m[31m.[39m[31m283[39m
#> [38;5;250m 5[39m 2017-01-05 -[31m0[39m[31m.[39m[31m554[39m
#> [38;5;250m 6[39m 2017-01-06  0.629
#> [38;5;250m 7[39m 2017-01-07  2.07 
#> [38;5;250m 8[39m 2017-01-08 -[31m1[39m[31m.[39m[31m63[39m 
#> [38;5;250m 9[39m 2017-01-09  0.512
#> [38;5;250m10[39m 2017-01-10 -[31m1[39m[31m.[39m[31m86[39m 

# specify the index var
as_tsibble(tbl1, index = date)

#> [38;5;246m# A tsibble: 10 x 2 [1D][39m
#>    date        value
#>    [3m[38;5;246m<date>[39m[23m      [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m 1[39m 2017-01-01 -[31m1[39m[31m.[39m[31m82[39m 
#> [38;5;250m 2[39m 2017-01-02 -[31m0[39m[31m.[39m[31m247[39m
#> [38;5;250m 3[39m 2017-01-03 -[31m0[39m[31m.[39m[31m244[39m
#> [38;5;250m 4[39m 2017-01-04 -[31m0[39m[31m.[39m[31m283[39m
#> [38;5;250m 5[39m 2017-01-05 -[31m0[39m[31m.[39m[31m554[39m
#> [38;5;250m 6[39m 2017-01-06  0.629
#> [38;5;250m 7[39m 2017-01-07  2.07 
#> [38;5;250m 8[39m 2017-01-08 -[31m1[39m[31m.[39m[31m63[39m 
#> [38;5;250m 9[39m 2017-01-09  0.512
#> [38;5;250m10[39m 2017-01-10 -[31m1[39m[31m.[39m[31m86[39m 

# coerce tibble to tsibble with one key ----
# "date" is automatically considered as the index var, and "group" is the key
tbl2 <- tibble(
  mth = rep(yearmonth("201701") + 0:9, 3),
  group = rep(c("x", "y", "z"), each = 10),
  value = rnorm(30)
)
as_tsibble(tbl2, key = id(group))

#> Column `mth` is the index.

#> [38;5;246m# A tsibble: 30 x 3 [1M][39m
#> [38;5;246m# Key:       group [3][39m
#>         mth group   value
#>       [3m[38;5;246m<mth>[39m[23m [3m[38;5;246m<chr>[39m[23m   [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m 1[39m 2017 Jan x     -[31m0[39m[31m.[39m[31m522[39m 
#> [38;5;250m 2[39m 2017 Feb x     -[31m0[39m[31m.[39m[31m0[39m[31m52[4m6[24m[39m
#> [38;5;250m 3[39m 2017 Mar x      0.543 
#> [38;5;250m 4[39m 2017 Apr x     -[31m0[39m[31m.[39m[31m914[39m 
#> [38;5;250m 5[39m 2017 May x      0.468 
#> [38;5;250m 6[39m 2017 Jun x      0.363 
#> [38;5;250m 7[39m 2017 Jul x     -[31m1[39m[31m.[39m[31m30[39m  
#> [38;5;250m 8[39m 2017 Aug x      0.738 
#> [38;5;250m 9[39m 2017 Sep x      1.89  
#> [38;5;250m10[39m 2017 Oct x     -[31m0[39m[31m.[39m[31m0[39m[31m97[4m4[24m[39m
#> [38;5;246m# ... with 20 more rows[39m

as_tsibble(tbl2, key = id(group), index = mth)

#> [38;5;246m# A tsibble: 30 x 3 [1M][39m
#> [38;5;246m# Key:       group [3][39m
#>         mth group   value
#>       [3m[38;5;246m<mth>[39m[23m [3m[38;5;246m<chr>[39m[23m   [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m 1[39m 2017 Jan x     -[31m0[39m[31m.[39m[31m522[39m 
#> [38;5;250m 2[39m 2017 Feb x     -[31m0[39m[31m.[39m[31m0[39m[31m52[4m6[24m[39m
#> [38;5;250m 3[39m 2017 Mar x      0.543 
#> [38;5;250m 4[39m 2017 Apr x     -[31m0[39m[31m.[39m[31m914[39m 
#> [38;5;250m 5[39m 2017 May x      0.468 
#> [38;5;250m 6[39m 2017 Jun x      0.363 
#> [38;5;250m 7[39m 2017 Jul x     -[31m1[39m[31m.[39m[31m30[39m  
#> [38;5;250m 8[39m 2017 Aug x      0.738 
#> [38;5;250m 9[39m 2017 Sep x      1.89  
#> [38;5;250m10[39m 2017 Oct x     -[31m0[39m[31m.[39m[31m0[39m[31m97[4m4[24m[39m
#> [38;5;246m# ... with 20 more rows[39m

# coerce ts to tsibble
as_tsibble(AirPassengers)

#> [38;5;246m# A tsibble: 144 x 2 [1M][39m
#>       index value
#>       [3m[38;5;246m<mth>[39m[23m [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m 1[39m 1949 Jan   112
#> [38;5;250m 2[39m 1949 Feb   118
#> [38;5;250m 3[39m 1949 Mar   132
#> [38;5;250m 4[39m 1949 Apr   129
#> [38;5;250m 5[39m 1949 May   121
#> [38;5;250m 6[39m 1949 Jun   135
#> [38;5;250m 7[39m 1949 Jul   148
#> [38;5;250m 8[39m 1949 Aug   148
#> [38;5;250m 9[39m 1949 Sep   136
#> [38;5;250m10[39m 1949 Oct   119
#> [38;5;246m# ... with 134 more rows[39m

as_tsibble(sunspot.year)

#> [38;5;246m# A tsibble: 289 x 2 [1Y][39m
#>    index value
#>    [3m[38;5;246m<dbl>[39m[23m [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m 1[39m  [4m1[24m700     5
#> [38;5;250m 2[39m  [4m1[24m701    11
#> [38;5;250m 3[39m  [4m1[24m702    16
#> [38;5;250m 4[39m  [4m1[24m703    23
#> [38;5;250m 5[39m  [4m1[24m704    36
#> [38;5;250m 6[39m  [4m1[24m705    58
#> [38;5;250m 7[39m  [4m1[24m706    29
#> [38;5;250m 8[39m  [4m1[24m707    20
#> [38;5;250m 9[39m  [4m1[24m708    10
#> [38;5;250m10[39m  [4m1[24m709     8
#> [38;5;246m# ... with 279 more rows[39m

as_tsibble(sunspot.month)

#> [38;5;246m# A tsibble: 3,177 x 2 [1M][39m
#>       index value
#>       [3m[38;5;246m<mth>[39m[23m [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m 1[39m 1749 Jan  58  
#> [38;5;250m 2[39m 1749 Feb  62.6
#> [38;5;250m 3[39m 1749 Mar  70  
#> [38;5;250m 4[39m 1749 Apr  55.7
#> [38;5;250m 5[39m 1749 May  85  
#> [38;5;250m 6[39m 1749 Jun  83.5
#> [38;5;250m 7[39m 1749 Jul  94.8
#> [38;5;250m 8[39m 1749 Aug  66.3
#> [38;5;250m 9[39m 1749 Sep  75.9
#> [38;5;250m10[39m 1749 Oct  75.5
#> [38;5;246m# ... with 3,167 more rows[39m

as_tsibble(austres)

#> [38;5;246m# A tsibble: 89 x 2 [1Q][39m
#>      index  value
#>      [3m[38;5;246m<qtr>[39m[23m  [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m 1[39m 1971 Q2 [4m1[24m[4m3[24m067.
#> [38;5;250m 2[39m 1971 Q3 [4m1[24m[4m3[24m130.
#> [38;5;250m 3[39m 1971 Q4 [4m1[24m[4m3[24m198.
#> [38;5;250m 4[39m 1972 Q1 [4m1[24m[4m3[24m254.
#> [38;5;250m 5[39m 1972 Q2 [4m1[24m[4m3[24m304.
#> [38;5;250m 6[39m 1972 Q3 [4m1[24m[4m3[24m354.
#> [38;5;250m 7[39m 1972 Q4 [4m1[24m[4m3[24m409.
#> [38;5;250m 8[39m 1973 Q1 [4m1[24m[4m3[24m459.
#> [38;5;250m 9[39m 1973 Q2 [4m1[24m[4m3[24m504.
#> [38;5;250m10[39m 1973 Q3 [4m1[24m[4m3[24m553.
#> [38;5;246m# ... with 79 more rows[39m

# coerce mts to tsibble
z <- ts(matrix(rnorm(300), 100, 3), start = c(1961, 1), frequency = 12)
as_tsibble(z)

#> [38;5;246m# A tsibble: 300 x 3 [1M][39m
#> [38;5;246m# Key:       key [3][39m
#>       index key        value
#>       [3m[38;5;246m<mth>[39m[23m [3m[38;5;246m<chr>[39m[23m      [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m 1[39m 1961 Jan Series 1  0.046[4m5[24m
#> [38;5;250m 2[39m 1961 Feb Series 1  0.578 
#> [38;5;250m 3[39m 1961 Mar Series 1  0.118 
#> [38;5;250m 4[39m 1961 Apr Series 1 -[31m1[39m[31m.[39m[31m91[39m  
#> [38;5;250m 5[39m 1961 May Series 1  0.862 
#> [38;5;250m 6[39m 1961 Jun Series 1 -[31m0[39m[31m.[39m[31m243[39m 
#> [38;5;250m 7[39m 1961 Jul Series 1 -[31m0[39m[31m.[39m[31m206[39m 
#> [38;5;250m 8[39m 1961 Aug Series 1  0.019[4m2[24m
#> [38;5;250m 9[39m 1961 Sep Series 1  0.029[4m6[24m
#> [38;5;250m10[39m 1961 Oct Series 1  0.550 
#> [38;5;246m# ... with 290 more rows[39m

as_tsibble(z, gather = FALSE)

#> [38;5;246m# A tsibble: 100 x 4 [1M][39m
#>       index `Series 1` `Series 2` `Series 3`
#>       [3m[38;5;246m<mth>[39m[23m      [3m[38;5;246m<dbl>[39m[23m      [3m[38;5;246m<dbl>[39m[23m      [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m 1[39m 1961 Jan     0.046[4m5[24m     -[31m1[39m[31m.[39m[31m12[39m     0.003[4m1[24m[4m3[24m
#> [38;5;250m 2[39m 1961 Feb     0.578       0.515   -[31m0[39m[31m.[39m[31m0[39m[31m77[4m9[24m[39m 
#> [38;5;250m 3[39m 1961 Mar     0.118      -[31m1[39m[31m.[39m[31m51[39m     0.441  
#> [38;5;250m 4[39m 1961 Apr    -[31m1[39m[31m.[39m[31m91[39m        1.53     0.129  
#> [38;5;250m 5[39m 1961 May     0.862       0.429   -[31m0[39m[31m.[39m[31m830[39m  
#> [38;5;250m 6[39m 1961 Jun    -[31m0[39m[31m.[39m[31m243[39m       0.122   -[31m0[39m[31m.[39m[31m504[39m  
#> [38;5;250m 7[39m 1961 Jul    -[31m0[39m[31m.[39m[31m206[39m      -[31m1[39m[31m.[39m[31m14[39m    -[31m1[39m[31m.[39m[31m19[39m   
#> [38;5;250m 8[39m 1961 Aug     0.019[4m2[24m     -[31m0[39m[31m.[39m[31m558[39m   -[31m0[39m[31m.[39m[31m752[39m  
#> [38;5;250m 9[39m 1961 Sep     0.029[4m6[24m      1.05     1.46   
#> [38;5;250m10[39m 1961 Oct     0.550       0.678   -[31m0[39m[31m.[39m[31m829[39m  
#> [38;5;246m# ... with 90 more rows[39m

# coerce hts from the "hts" package to tsibble
if (!requireNamespace("hts", quietly = TRUE)) {
  stop("Please install the hts package to run these following examples.")
}
as_tsibble(hts::htseg1)

#> [38;5;246m# A tsibble: 50 x 4 [1Y][39m
#> [38;5;246m# Key:       Level 1, Level 2 [5][39m
#>    index value `Level 1` `Level 2`
#>    [3m[38;5;246m<dbl>[39m[23m [3m[38;5;246m<dbl>[39m[23m [3m[38;5;246m<chr>[39m[23m     [3m[38;5;246m<chr>[39m[23m    
#> [38;5;250m 1[39m  [4m1[24m992  7.86 A         AA       
#> [38;5;250m 2[39m  [4m1[24m993  8.34 A         AA       
#> [38;5;250m 3[39m  [4m1[24m994  8.55 A         AA       
#> [38;5;250m 4[39m  [4m1[24m995  8.60 A         AA       
#> [38;5;250m 5[39m  [4m1[24m996  8.77 A         AA       
#> [38;5;250m 6[39m  [4m1[24m997  8.84 A         AA       
#> [38;5;250m 7[39m  [4m1[24m998  8.89 A         AA       
#> [38;5;250m 8[39m  [4m1[24m999  8.93 A         AA       
#> [38;5;250m 9[39m  [4m2[24m000  9.13 A         AA       
#> [38;5;250m10[39m  [4m2[24m001  9.15 A         AA       
#> [38;5;246m# ... with 40 more rows[39m

as_tsibble(hts::htseg2)

#> [38;5;246m# A tsibble: 160 x 5 [1Y][39m
#> [38;5;246m# Key:       Level 1, Level 2, Level 3 [10][39m
#>    index value `Level 1` `Level 2` `Level 3`
#>    [3m[38;5;246m<dbl>[39m[23m [3m[38;5;246m<dbl>[39m[23m [3m[38;5;246m<chr>[39m[23m     [3m[38;5;246m<chr>[39m[23m     [3m[38;5;246m<chr>[39m[23m    
#> [38;5;250m 1[39m  [4m1[24m992 -[31m2[39m[31m.[39m[31m42[39m A         A10       A10A     
#> [38;5;250m 2[39m  [4m1[24m993 -[31m2[39m[31m.[39m[31m20[39m A         A10       A10A     
#> [38;5;250m 3[39m  [4m1[24m994 -[31m2[39m[31m.[39m[31m19[39m A         A10       A10A     
#> [38;5;250m 4[39m  [4m1[24m995 -[31m1[39m[31m.[39m[31m93[39m A         A10       A10A     
#> [38;5;250m 5[39m  [4m1[24m996 -[31m1[39m[31m.[39m[31m64[39m A         A10       A10A     
#> [38;5;250m 6[39m  [4m1[24m997 -[31m1[39m[31m.[39m[31m63[39m A         A10       A10A     
#> [38;5;250m 7[39m  [4m1[24m998 -[31m1[39m[31m.[39m[31m61[39m A         A10       A10A     
#> [38;5;250m 8[39m  [4m1[24m999 -[31m1[39m[31m.[39m[31m59[39m A         A10       A10A     
#> [38;5;250m 9[39m  [4m2[24m000 -[31m1[39m[31m.[39m[31m53[39m A         A10       A10A     
#> [38;5;250m10[39m  [4m2[24m001 -[31m1[39m[31m.[39m[31m39[39m A         A10       A10A     
#> [38;5;246m# ... with 150 more rows[39m