Deprecated functions

A thin wrapper of dplyr::case_when() if there are NAs

Find which row has duplicated key and index elements

split_by(x, ...)

case_na(formula)

as.tsibble(x, ...)

fill_na(.data, ..., .full = FALSE)

find_duplicates(data, key = id(), index, fromLast = FALSE)

Arguments

x
A data frame.

...
A list of unquoted variables, separated by commas, to split a dataset.

formula
A two-sided formula. The LHS expects a vector containing NA, and the RHS gives the replacement value.

data
A tbl_ts object.

key
Structural variable(s) that define unique time indices, used with the helper id. If a univariate time series (without an explicit key), simply call id().

index
A bare (or unquoted) variable to specify the time index variable.

fromLast
TRUE does the duplication check from the last of identical elements.

Value

A logical vector of the same length as the row number of data

See also

Examples

pedestrian %>%
  split_by(Sensor)

#> Warning: This function will be defunct soon.
#> Please use `split()` instead.

#> [[1]]
#> # A tsibble: 14,566 x 5 [1h] <Australia/Melbourne>
#> # Key:       Sensor [1]
#>    Sensor         Date_Time           Date        Time Count
#>    <chr>          <dttm>              <date>     <int> <int>
#>  1 Birrarung Marr 2015-01-01 00:00:00 2015-01-01     0  1630
#>  2 Birrarung Marr 2015-01-01 01:00:00 2015-01-01     1   826
#>  3 Birrarung Marr 2015-01-01 02:00:00 2015-01-01     2   567
#>  4 Birrarung Marr 2015-01-01 03:00:00 2015-01-01     3   264
#>  5 Birrarung Marr 2015-01-01 04:00:00 2015-01-01     4   139
#>  6 Birrarung Marr 2015-01-01 05:00:00 2015-01-01     5    77
#>  7 Birrarung Marr 2015-01-01 06:00:00 2015-01-01     6    44
#>  8 Birrarung Marr 2015-01-01 07:00:00 2015-01-01     7    56
#>  9 Birrarung Marr 2015-01-01 08:00:00 2015-01-01     8   113
#> 10 Birrarung Marr 2015-01-01 09:00:00 2015-01-01     9   166
#> # … with 14,556 more rows
#> 
#> [[2]]
#> # A tsibble: 16,414 x 5 [1h] <Australia/Melbourne>
#> # Key:       Sensor [1]
#>    Sensor                     Date_Time           Date        Time Count
#>    <chr>                      <dttm>              <date>     <int> <int>
#>  1 Bourke Street Mall (North) 2015-02-17 00:00:00 2015-02-17     0    61
#>  2 Bourke Street Mall (North) 2015-02-17 01:00:00 2015-02-17     1    16
#>  3 Bourke Street Mall (North) 2015-02-17 02:00:00 2015-02-17     2    15
#>  4 Bourke Street Mall (North) 2015-02-17 03:00:00 2015-02-17     3    17
#>  5 Bourke Street Mall (North) 2015-02-17 04:00:00 2015-02-17     4     9
#>  6 Bourke Street Mall (North) 2015-02-17 05:00:00 2015-02-17     5    25
#>  7 Bourke Street Mall (North) 2015-02-17 06:00:00 2015-02-17     6    88
#>  8 Bourke Street Mall (North) 2015-02-17 07:00:00 2015-02-17     7   279
#>  9 Bourke Street Mall (North) 2015-02-17 08:00:00 2015-02-17     8   770
#> 10 Bourke Street Mall (North) 2015-02-17 09:00:00 2015-02-17     9   722
#> # … with 16,404 more rows
#> 
#> [[3]]
#> # A tsibble: 17,518 x 5 [1h] <Australia/Melbourne>
#> # Key:       Sensor [1]
#>    Sensor                        Date_Time           Date        Time Count
#>    <chr>                         <dttm>              <date>     <int> <int>
#>  1 QV Market-Elizabeth St (West) 2015-01-01 00:00:00 2015-01-01     0   490
#>  2 QV Market-Elizabeth St (West) 2015-01-01 01:00:00 2015-01-01     1   397
#>  3 QV Market-Elizabeth St (West) 2015-01-01 02:00:00 2015-01-01     2   247
#>  4 QV Market-Elizabeth St (West) 2015-01-01 03:00:00 2015-01-01     3   172
#>  5 QV Market-Elizabeth St (West) 2015-01-01 04:00:00 2015-01-01     4   128
#>  6 QV Market-Elizabeth St (West) 2015-01-01 05:00:00 2015-01-01     5    66
#>  7 QV Market-Elizabeth St (West) 2015-01-01 06:00:00 2015-01-01     6    72
#>  8 QV Market-Elizabeth St (West) 2015-01-01 07:00:00 2015-01-01     7    69
#>  9 QV Market-Elizabeth St (West) 2015-01-01 08:00:00 2015-01-01     8   106
#> 10 QV Market-Elizabeth St (West) 2015-01-01 09:00:00 2015-01-01     9   217
#> # … with 17,508 more rows
#> 
#> [[4]]
#> # A tsibble: 17,539 x 5 [1h] <Australia/Melbourne>
#> # Key:       Sensor [1]
#>    Sensor                 Date_Time           Date        Time Count
#>    <chr>                  <dttm>              <date>     <int> <int>
#>  1 Southern Cross Station 2015-01-01 00:00:00 2015-01-01     0   746
#>  2 Southern Cross Station 2015-01-01 01:00:00 2015-01-01     1   312
#>  3 Southern Cross Station 2015-01-01 02:00:00 2015-01-01     2   180
#>  4 Southern Cross Station 2015-01-01 03:00:00 2015-01-01     3   133
#>  5 Southern Cross Station 2015-01-01 04:00:00 2015-01-01     4    44
#>  6 Southern Cross Station 2015-01-01 05:00:00 2015-01-01     5    16
#>  7 Southern Cross Station 2015-01-01 06:00:00 2015-01-01     6    13
#>  8 Southern Cross Station 2015-01-01 07:00:00 2015-01-01     7    21
#>  9 Southern Cross Station 2015-01-01 08:00:00 2015-01-01     8    39
#> 10 Southern Cross Station 2015-01-01 09:00:00 2015-01-01     9    36
#> # … with 17,529 more rows
#> 

x <- rnorm(10)
x[c(3, 7)] <- NA_real_
case_na(x ~ 10)

#> Warning: This function will be defunct soon.

#>  [1] -0.2343379  2.0866886 10.0000000 -1.3928471 -1.1422908  1.7046087
#>  [7] 10.0000000 -0.4372812 -0.1192151  0.7864629

case_na(x ~ mean(x, na.rm = TRUE))

#> Warning: This function will be defunct soon.

#>  [1] -0.2343379  2.0866886  0.1564735 -1.3928471 -1.1422908  1.7046087
#>  [7]  0.1564735 -0.4372812 -0.1192151  0.7864629