Find duplication of key and index variables

Find which row has duplicated key and index elements

find_duplicates(data, key = id(), index, fromLast = FALSE)

Arguments

data
A tbl_ts object.

key
Structural variable(s) that define unique time indices, used with the helper id. If a univariate time series (without an explicit key), simply call id().

index
A bare (or unquoted) variable to specify the time index variable.

fromLast
TRUE does the duplication check from the last of identical elements.

Value

A logical vector of the same length as the row number of data