Gather columns into key-value pairs.

Gather columns into key-value pairs.

# S3 method for tbl_ts
gather(data, key = "key", value = "value", ...,
  na.rm = FALSE, convert = FALSE, factor_key = FALSE)

Arguments

data
A tbl_ts.

key
Names of new key and value columns, as strings or symbols.

This argument is passed by expression and supports quasiquotation (you can unquote strings and symbols). The name is captured from the expression with rlang::quo_name() (note that this kind of interface where symbols do not represent actual objects is now discouraged in the tidyverse; we support it here for backward compatibility).

value
Names of new key and value columns, as strings or symbols.

This argument is passed by expression and supports quasiquotation (you can unquote strings and symbols). The name is captured from the expression with rlang::quo_name() (note that this kind of interface where symbols do not represent actual objects is now discouraged in the tidyverse; we support it here for backward compatibility).

...
A selection of columns. If empty, all variables are selected. You can supply bare variable names, select all variables between x and z with x:z, exclude y with -y. For more options, see the dplyr::select() documentation. See also the section on selection rules below.

na.rm
If TRUE, will remove rows from output where the value column in NA.

convert
If TRUE will automatically run type.convert() on the key column. This is useful if the column names are actually numeric, integer, or logical.

factor_key
If FALSE, the default, the key values will be stored as a character vector. If TRUE, will be stored as a factor, which preserves the original ordering of the columns.

See also

Examples

# example from tidyr
stocks <- tsibble(
  time = as.Date('2009-01-01') + 0:9,
  X = rnorm(10, 0, 1),
  Y = rnorm(10, 0, 2),
  Z = rnorm(10, 0, 4)
)

#> The `index` is `time`.

stocks %>% gather(stock, price, -time)

#> # A tsibble: 30 x 3 [1DAY]
#> # Keys:      stock [3]
#>    time       stock  price
#>    <date>     <chr>  <dbl>
#>  1 2009-01-01 X     -0.384
#>  2 2009-01-02 X      1.83 
#>  3 2009-01-03 X     -0.551
#>  4 2009-01-04 X     -0.866
#>  5 2009-01-05 X     -0.344
#>  6 2009-01-06 X      1.06 
#>  7 2009-01-07 X      0.813
#>  8 2009-01-08 X      1.80 
#>  9 2009-01-09 X     -0.105
#> 10 2009-01-10 X      0.982
#> # ... with 20 more rows