Group by one or more variables

Group by one or more variables

# S3 method for tbl_ts
group_by(.data, ..., add = FALSE)

# S3 method for grouped_ts
ungroup(x, ...)

Arguments

.data
A tsibble.

...
Variables to group by. It follows a consistent rule as the "key" expression in as_tsibble, which means | for nested variables and , for crossed variables. The following operations will affect the tsibble structure based on the way how the variables are grouped.

add
TRUE adds to the existing groups, otherwise overwrites.

x
A (grouped) tsibble.

See also

dplyr::group_by dplyr::ungroup

Examples

data(tourism)
tourism %>%
  group_by(Region | State) %>%
  summarise(geo_trips = sum(Trips))

#> # A tsibble: 5,852 x 4 [1QUARTER]
#> # Keys: Region | State [77]
#> # Groups: Region | State [77]
#>    Quarter Region                     State              geo_trips
#>      <qtr> <chr>                      <chr>                  <dbl>
#>  1 1998 Q1 Adelaide                   South Australia       659   
#>  2 1998 Q1 Adelaide Hills             South Australia         9.80
#>  3 1998 Q1 Alice Springs              Northern Territory     20.2 
#>  4 1998 Q1 Australia's Coral Coast    Western Australia     133   
#>  5 1998 Q1 Australia's Golden Outback Western Australia     162   
#>  6 1998 Q1 Australia's North West     Western Australia     121   
#>  7 1998 Q1 Australia's South West     Western Australia     475   
#>  8 1998 Q1 Ballarat                   Victoria              182   
#>  9 1998 Q1 Barkly                     Northern Territory     18.5 
#> 10 1998 Q1 Barossa                    South Australia        46.8 
#> # ... with 5,842 more rows