Group by key variables

Group by key variables

group_by_key(.tbl, .funs = list(), ...)

Arguments

.tbl
A tbl_ts object.

.funs
List of function calls generated by funs(), or a character vector of function names, or simply a function.

Bare formulas are passed to rlang::as_function() to create purrr-style lambda functions. Note that these lambda prevent hybrid evaluation from happening and it is thus more efficient to supply functions like mean() directly rather than in a lambda-formula.

...
Additional arguments for the function calls in .funs. These are evaluated only once, with tidy dots support.

Examples

tourism %>%
  group_by_key()

#> # A tsibble: 23,408 x 5 [1Q]
#> # Key:       Region | State, Purpose [308]
#> # Groups:    Region, State, Purpose [308]
#>    Quarter Region   State           Purpose  Trips
#>  *   <qtr> <chr>    <chr>           <chr>    <dbl>
#>  1 1998 Q1 Adelaide South Australia Business  135.
#>  2 1998 Q2 Adelaide South Australia Business  110.
#>  3 1998 Q3 Adelaide South Australia Business  166.
#>  4 1998 Q4 Adelaide South Australia Business  127.
#>  5 1999 Q1 Adelaide South Australia Business  137.
#>  6 1999 Q2 Adelaide South Australia Business  200.
#>  7 1999 Q3 Adelaide South Australia Business  169.
#>  8 1999 Q4 Adelaide South Australia Business  134.
#>  9 2000 Q1 Adelaide South Australia Business  154.
#> 10 2000 Q2 Adelaide South Australia Business  169.
#> # ... with 23,398 more rows