Does a tsibble have implicit gaps in time?

Does a tsibble have implicit gaps in time?

has_gaps(.data, .full = FALSE, ...)

Arguments

.data
A tbl_ts.

.full
FALSE to find gaps for each series within its own period. TRUE to find gaps over the entire time span of the data.

...
Other arguments passed on to individual methods.

Value

A tibble contains "key" variables and new column .gaps of TRUE/FALSE.

See also

Examples

harvest <- tsibble(
  year = c(2010, 2011, 2013, 2011, 2012, 2013),
  fruit = rep(c("kiwi", "cherry"), each = 3),
  kilo = sample(1:10, size = 6),
<<<<<<< HEAD
  key = id(fruit), index = year
=======
  key = fruit, index = year
>>>>>>> 1acf3859749ceb61f22a7933b0fe557e9d6ca4ce
)
has_gaps(harvest)

#> # A tibble: 2 x 2
#>   fruit  .gaps
#>   <chr>  <lgl>
#> 1 cherry FALSE
#> 2 kiwi   TRUE 

has_gaps(harvest, .full = TRUE)

#> # A tibble: 2 x 2
#>   fruit  .gaps
#>   <chr>  <lgl>
#> 1 cherry TRUE 
#> 2 kiwi   TRUE