Change/update key variables for a given tbl_ts

Change/update key variables for a given tbl_ts

key_update(.data, ..., validate = TRUE)

Arguments

.data
A tbl_ts.

...
Expressions used to construct the key:

unspecified: drop every single variable from the old key.

| and , for nesting and crossing factors.

validate
TRUE suggests to verify that each key or each combination of key variables lead to unique time indices (i.e. a valid tsibble). It will also make sure that the nested variables are arranged from lower level to higher, if nested variables are passed to key. If you are sure that it's a valid input, specify FALSE to skip the checks.

Examples

# tourism could be identified by Region and Purpose ----
tourism %>%
  key_update(Region, Purpose)

#> # A tsibble: 23,408 x 5 [1Q]
#> # Key:       Region, Purpose [308]
#>    Quarter Region   State           Purpose  Trips
#>  *   <qtr> <chr>    <chr>           <chr>    <dbl>
#>  1 1998 Q1 Adelaide South Australia Business  135.
#>  2 1998 Q2 Adelaide South Australia Business  110.
#>  3 1998 Q3 Adelaide South Australia Business  166.
#>  4 1998 Q4 Adelaide South Australia Business  127.
#>  5 1999 Q1 Adelaide South Australia Business  137.
#>  6 1999 Q2 Adelaide South Australia Business  200.
#>  7 1999 Q3 Adelaide South Australia Business  169.
#>  8 1999 Q4 Adelaide South Australia Business  134.
#>  9 2000 Q1 Adelaide South Australia Business  154.
#> 10 2000 Q2 Adelaide South Australia Business  169.
#> # ... with 23,398 more rows