Tsibble-aware lag, lead, and difference operations

They are specialist vector functions, used in tsibble data context inside mutate()/transmute(). Unlike their counterparts, such as lag() and lead(), keyed_*() take care of temporal ordering and gaps, and do the right thing. They ignore the grouping data structure.

keyed_lag(select, n = 1L, default = NA)

keyed_lead(select, n = 1L, default = NA)

keyed_difference(select, lag = 1, differences = 1, default = NA)

Arguments

select
Unquoted variable.

n
a positive integer of length 1, giving the number of positions to lead or lag by

default
value used for non-existent rows. Defaults to NA.

lag
A positive integer indicating which lag to use.

differences
A positive integer indicating the order of the difference.

Examples

library(dplyr)
yrmth <- c("2018-07", "2018-08", "2018-11", "2018-12", "2019-01", "2019-03")
tsbl <- tsibble(
  yrmth = yearmonth(yrmth),
  income = c(1153, 1181, 1236, 1297, 1264, 1282),
  index = yrmth)

tsbl %>%
  mutate(
    lag_income = keyed_lag(income),
    lead_income = keyed_lead(income),
    diff_income = keyed_difference(income)
  )

#> # A tsibble: 6 x 5 [1M]
#>      yrmth income lag_income lead_income diff_income
#>      <mth>  <dbl>      <dbl>       <dbl>       <dbl>
#> 1 2018 Jul   1153         NA        1181          NA
#> 2 2018 Aug   1181       1153          NA         -28
#> 3 2018 Nov   1236         NA        1297          NA
#> 4 2018 Dec   1297       1236        1264         -61
#> 5 2019 Jan   1264       1297          NA          33
#> 6 2019 Mar   1282         NA          NA          NA

# take care of key and gaps
pedestrian %>%
  mutate(l_count = keyed_lag(Count))

#> # A tsibble: 66,037 x 6 [1h] <Australia/Melbourne>
#> # Key:       Sensor [4]
#>    Sensor         Date_Time           Date        Time Count l_count
#>    <chr>          <dttm>              <date>     <int> <int>   <int>
#>  1 Birrarung Marr 2015-01-01 00:00:00 2015-01-01     0  1630      NA
#>  2 Birrarung Marr 2015-01-01 01:00:00 2015-01-01     1   826    1630
#>  3 Birrarung Marr 2015-01-01 02:00:00 2015-01-01     2   567     826
#>  4 Birrarung Marr 2015-01-01 03:00:00 2015-01-01     3   264     567
#>  5 Birrarung Marr 2015-01-01 04:00:00 2015-01-01     4   139     264
#>  6 Birrarung Marr 2015-01-01 05:00:00 2015-01-01     5    77     139
#>  7 Birrarung Marr 2015-01-01 06:00:00 2015-01-01     6    44      77
#>  8 Birrarung Marr 2015-01-01 07:00:00 2015-01-01     7    56      44
#>  9 Birrarung Marr 2015-01-01 08:00:00 2015-01-01     8   113      56
#> 10 Birrarung Marr 2015-01-01 09:00:00 2015-01-01     9   166     113
#> # … with 66,027 more rows

# to replace the following chain
pedestrian %>%
  fill_gaps() %>%
  group_by_key() %>%
  mutate(l_count = lag(Count))

#> # A tsibble: 69,048 x 6 [1h] <Australia/Melbourne>
#> # Key:       Sensor [4]
#> # Groups:    Sensor [4]
#>    Sensor         Date_Time           Date        Time Count l_count
#>    <chr>          <dttm>              <date>     <int> <int>   <int>
#>  1 Birrarung Marr 2015-01-01 00:00:00 2015-01-01     0  1630      NA
#>  2 Birrarung Marr 2015-01-01 01:00:00 2015-01-01     1   826    1630
#>  3 Birrarung Marr 2015-01-01 02:00:00 2015-01-01     2   567     826
#>  4 Birrarung Marr 2015-01-01 03:00:00 2015-01-01     3   264     567
#>  5 Birrarung Marr 2015-01-01 04:00:00 2015-01-01     4   139     264
#>  6 Birrarung Marr 2015-01-01 05:00:00 2015-01-01     5    77     139
#>  7 Birrarung Marr 2015-01-01 06:00:00 2015-01-01     6    44      77
#>  8 Birrarung Marr 2015-01-01 07:00:00 2015-01-01     7    56      44
#>  9 Birrarung Marr 2015-01-01 08:00:00 2015-01-01     8   113      56
#> 10 Birrarung Marr 2015-01-01 09:00:00 2015-01-01     9   166     113
#> # … with 69,038 more rows