Nest repeated values in a list-variable.

Nest repeated values in a list-variable.

# S3 method for tbl_ts
nest(data, ..., .key = "data")

Arguments

data
A tbl_ts.

...
A selection of columns. If empty, all variables are selected. You can supply bare variable names, select all variables between x and z with x:z, exclude y with -y. For more options, see the dplyr::select() documentation. See also the section on selection rules below.

.key
The name of the new column, as a string or symbol.

This argument is passed by expression and supports quasiquotation (you can unquote strings and symbols). The name is captured from the expression with rlang::quo_name() (note that this kind of interface where symbols do not represent actual objects is now discouraged in the tidyverse; we support it here for backward compatibility).

Value

A tibble containing a list column of tbl_ts.

See also

Examples

pedestrian %>%
  nest(-Sensor)

#> # A tibble: 4 x 2
#>   Sensor                        data                  
#>   <chr>                         <list>                
#> 1 Birrarung Marr                <tsibble [14,592 × 4]>
#> 2 Bourke Street Mall (North)    <tsibble [16,414 × 4]>
#> 3 QV Market-Elizabeth St (West) <tsibble [17,526 × 4]>
#> 4 Southern Cross Station        <tsibble [17,539 × 4]>

pedestrian %>%
  group_by(Sensor) %>%
  nest()

#> # A tibble: 4 x 2
#>   Sensor                        data                  
#>   <chr>                         <list>                
#> 1 Birrarung Marr                <tsibble [14,592 × 4]>
#> 2 Bourke Street Mall (North)    <tsibble [16,414 × 4]>
#> 3 QV Market-Elizabeth St (West) <tsibble [17,526 × 4]>
#> 4 Southern Cross Station        <tsibble [17,539 × 4]>