Tiling window calculation

Tiling window without overlapping observations:

tile() always returns a list.

tile_lgl(), tile_int(), tile_dbl(), tile_chr() use the same arguments as tile(), but return vectors of the corresponding type.

tile_dfr() tile_dfc() return data frames using row-binding & column-binding.

tile(.x, .f, ..., .size = 1)

tile_dfr(.x, .f, ..., .size = 1, .id = NULL)

tile_dfc(.x, .f, ..., .size = 1)

Arguments

.x
An object to slide over.

.f
A function, formula, or atomic vector.

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function. There are three ways to refer to the arguments:

For a single argument function, use .

For a two argument function, use .x and .y

For more arguments, use ..1, ..2, ..3 etc

This syntax allows you to create very compact anonymous functions.

If character vector, numeric vector, or list, it is converted to an extractor function. Character vectors index by name and numeric vectors index by position; use a list to index by position and name at different levels. Within a list, wrap strings in get-attr() to extract named attributes. If a component is not present, the value of .default will be returned.

...
Additional arguments passed on to .f.

.size
An integer for window size. If positive, moving forward from left to right; if negative, moving backward (from right to left).

.id
If not NULL a variable with this name will be created giving either the name or the index of the data frame.

See also

Examples

.x <- 1:5
.lst <- list(x = .x, y = 6:10, z = 11:15)
tile_dbl(.x, mean, .size = 2)

#> [1] 1.5 3.5 5.0

tile_lgl(.x, ~ mean(.) > 2, .size = 2)

#> [1] FALSE  TRUE  TRUE

tile(.lst, ~ ., .size = 2)

#> [[1]]
#> [[1]]$x
#> [1] 1 2 3 4 5
#> 
#> [[1]]$y
#> [1]  6  7  8  9 10
#> 
#> 
#> [[2]]
#> [[2]]$z
#> [1] 11 12 13 14 15
#> 
#>