Unnest a list column.

Unnest a list column.

# S3 method for lst_ts
unnest(data, ..., key = id(), .drop = NA, .id = NULL,
  .sep = NULL, .preserve = NULL)

Arguments

data
A lst_ts.

...
Specification of columns to nest. Use bare variable names or functions of variables. If omitted, defaults to all list-cols.

key
Unquoted variables to create the key (via id) after unnesting.

.drop
Should additional list columns be dropped? By default, unnest will drop them if unnesting the specified columns requires the rows to be duplicated.

.id
Data frame identifier - if supplied, will create a new column with name .id, giving a unique identifier. This is most useful if the list column is named.

.sep
If non-NULL, the names of unnested data frame columns will combine the name of the original list-col with the names from nested data frame, separated by .sep.

.preserve
Optionally, list-columns to preserve in the output. These will be duplicated in the same way as atomic vectors. This has dplyr::select semantics so you can preserve multiple variables with .preserve = c(x, y) or .preserve = starts_with("list").

Value

A tbl_ts.

See also

Examples

nested_ped <- pedestrian %>%
  nest(-Sensor)
nested_ped %>%
  unnest(key = id(Sensor))

#> # A tsibble: 66,071 x 5 [1HOUR]
#> # Keys:      Sensor [4]
#>    Sensor         Date_Time           Date        Time Count
#>    <chr>          <dttm>              <date>     <int> <int>
#>  1 Birrarung Marr 2015-01-01 00:00:00 2015-01-01     0  1630
#>  2 Birrarung Marr 2015-01-01 01:00:00 2015-01-01     1   826
#>  3 Birrarung Marr 2015-01-01 02:00:00 2015-01-01     2   567
#>  4 Birrarung Marr 2015-01-01 03:00:00 2015-01-01     3   264
#>  5 Birrarung Marr 2015-01-01 04:00:00 2015-01-01     4   139
#>  6 Birrarung Marr 2015-01-01 05:00:00 2015-01-01     5    77
#>  7 Birrarung Marr 2015-01-01 06:00:00 2015-01-01     6    44
#>  8 Birrarung Marr 2015-01-01 07:00:00 2015-01-01     7    56
#>  9 Birrarung Marr 2015-01-01 08:00:00 2015-01-01     8   113
#> 10 Birrarung Marr 2015-01-01 09:00:00 2015-01-01     9   166
#> # ... with 66,061 more rows

nested_tourism <- tourism %>%
  nest(-Region, -State)
nested_tourism %>%
  unnest(key = id(Region | State))

#> # A tsibble: 23,408 x 5 [1QUARTER]
#> # Keys:      Region | State, Purpose [308]
#>    Region   State           Quarter Purpose  Trips
#>    <chr>    <chr>             <qtr> <chr>    <dbl>
#>  1 Adelaide South Australia 1998 Q1 Business  135.
#>  2 Adelaide South Australia 1998 Q2 Business  110.
#>  3 Adelaide South Australia 1998 Q3 Business  166.
#>  4 Adelaide South Australia 1998 Q4 Business  127.
#>  5 Adelaide South Australia 1999 Q1 Business  137.
#>  6 Adelaide South Australia 1999 Q2 Business  200.
#>  7 Adelaide South Australia 1999 Q3 Business  169.
#>  8 Adelaide South Australia 1999 Q4 Business  134.
#>  9 Adelaide South Australia 2000 Q1 Business  154.
#> 10 Adelaide South Australia 2000 Q2 Business  169.
#> # ... with 23,398 more rows