Coerce to a tsibble object

Coerce to a tsibble object

as_tsibble(x, ...)

# S3 method for tbl_df
as_tsibble(x, key = id(), index, validate = TRUE,
  regular = TRUE)

# S3 method for data.frame
as_tsibble(x, key = id(), index, validate = TRUE,
  regular = TRUE)

# S3 method for list
as_tsibble(x, key = id(), index, validate = TRUE,
  regular = TRUE)

# S3 method for ts
as_tsibble(x, tz = "UTC", ...)

# S3 method for mts
as_tsibble(x, tz = "UTC", gather = TRUE, ...)

# S3 method for hts
as_tsibble(x, tz = "UTC", ...)

Arguments

x
Other objects to be coerced to a tsibble (tbl_ts).

...
Other arguments passed on to individual methods.

key
Structural variable(s) that define unique time indices, used with the helper id(). If a univariate time series (without an explicit key), simply call id().See below for details.

index
A bare (or unquoted) variable to specify the time index variable.

validate
TRUE suggests to verify that each key or each combination of key variables lead to unique time indices (i.e. a valid tsibble). It will also make sure that the nested variables are arranged from lower level to higher, if nested variables are passed to key. If you are sure that it's a valid input, specify FALSE to skip the checks.

regular
Regular time interval (TRUE) or irregular (FALSE). TRUE finds the minimal time span as the interval.

tz
Time zone. May be useful when a ts object is more frequent than daily.

gather
TRUE gives a "long" data form, otherwise as "wide" as x.

Value

A tsibble object.

Details

A valid tsibble does not arrange time in the chronological order. Please use arrange to get the order by time.

Index

The time indices are no longer an attribute (for example, the tsp attribute in a ts object), but preserved as the essential component of the tsibble. A few index class, such as Date, POSIXt, and difftime, forms the basis of the tsibble, with new additions yearmth and yearqtr representing year-month and year-quarter respectively.

Key

Key variable(s) together with the index uniquely identifies each record. And key(s) also imposes the structure on a tsibble to describe other measured variables:

None: an implicit variable

A single variable: an explicit variable. For example, data(pedestrian) uses the Sensor column as the key.

Nested variables: a nesting of one variable under another. For example, data(tourism) contains two geographical locations: Region and State. Region is the lower level than State in Australia; in other words, Region is nested into State, which naturally forms a hierarchy. A vertical bars (|) is used to describe this nesting relationship, and thus Region | State. In theory, nesting can involve infinite levels, so is tsibble.

Crossed variables: a crossing of one variable with another. For example, the geographical locations are crossed with the purpose of visiting (Purpose) in the data(tourism). A comma (,) is used to indicate this crossing relationship. Nested and crossed variables can be combined, such as data(tourism).

These key variables are descriptors, not values.

Interval

The interval function returns the interval associated with the tsibble.

Regular: the value and its time unit including "second", "minute", "hour", "day", "week", "month", "quarter", "year". An unrecognisable time interval is labelled as "unit".

Irregular: as_tsibble(regular = FALSE) gives the irregular tsibble. It is marked with "!".

Unknown: if there is only one entry for each key variable, the interval cannot be determined ("?").

Examples

# coerce tibble to tsibble w/o a key ----
tbl1 <- tibble::tibble(
  date = seq(as.Date("2017-01-01"), as.Date("2017-01-10"), by = 1),
  value = rnorm(10)
)
as_tsibble(tbl1)

#> The 'index' variable: date

#> # A tsibble: 10 x 2 [1DAY]
#>          date        value
#>  *     <date>        <dbl>
#>  1 2017-01-01 -1.400043517
#>  2 2017-01-02  0.255317055
#>  3 2017-01-03 -2.437263611
#>  4 2017-01-04 -0.005571287
#>  5 2017-01-05  0.621552721
#>  6 2017-01-06  1.148411606
#>  7 2017-01-07 -1.821817661
#>  8 2017-01-08 -0.247325302
#>  9 2017-01-09 -0.244199607
#> 10 2017-01-10 -0.282705449

# specify the index var
as_tsibble(tbl1, index = date)

#> # A tsibble: 10 x 2 [1DAY]
#>          date        value
#>  *     <date>        <dbl>
#>  1 2017-01-01 -1.400043517
#>  2 2017-01-02  0.255317055
#>  3 2017-01-03 -2.437263611
#>  4 2017-01-04 -0.005571287
#>  5 2017-01-05  0.621552721
#>  6 2017-01-06  1.148411606
#>  7 2017-01-07 -1.821817661
#>  8 2017-01-08 -0.247325302
#>  9 2017-01-09 -0.244199607
#> 10 2017-01-10 -0.282705449

# coerce tibble to tsibble with one key ----
# "date" is automatically considered as the index var, and "group" is the key
tbl2 <- tibble::tibble(
  date = rep(seq(as.Date("2017-01-01"), as.Date("2017-01-10"), by = 1), 3),
  group = rep(c("x", "y", "z"), each = 10),
  value = rnorm(30)
)
as_tsibble(tbl2, key = id(group))

#> The 'index' variable: date

#> # A tsibble: 30 x 3 [1DAY]
#> # Keys:      group
#>          date group       value
#>  *     <date> <chr>       <dbl>
#>  1 2017-01-01     x -0.55369938
#>  2 2017-01-02     x  0.62898204
#>  3 2017-01-03     x  2.06502490
#>  4 2017-01-04     x -1.63098940
#>  5 2017-01-05     x  0.51242695
#>  6 2017-01-06     x -1.86301149
#>  7 2017-01-07     x -0.52201251
#>  8 2017-01-08     x -0.05260191
#>  9 2017-01-09     x  0.54299634
#> 10 2017-01-10     x -0.91407483
#> # ... with 20 more rows

as_tsibble(tbl2, key = id(group), index = date)

#> # A tsibble: 30 x 3 [1DAY]
#> # Keys:      group
#>          date group       value
#>  *     <date> <chr>       <dbl>
#>  1 2017-01-01     x -0.55369938
#>  2 2017-01-02     x  0.62898204
#>  3 2017-01-03     x  2.06502490
#>  4 2017-01-04     x -1.63098940
#>  5 2017-01-05     x  0.51242695
#>  6 2017-01-06     x -1.86301149
#>  7 2017-01-07     x -0.52201251
#>  8 2017-01-08     x -0.05260191
#>  9 2017-01-09     x  0.54299634
#> 10 2017-01-10     x -0.91407483
#> # ... with 20 more rows

# coerce ts to tsibble
as_tsibble(AirPassengers)

#> # A tsibble: 144 x 2 [1MONTH]
#>       index value
#>  *    <mth> <dbl>
#>  1 1949 Jan   112
#>  2 1949 Feb   118
#>  3 1949 Mar   132
#>  4 1949 Apr   129
#>  5 1949 May   121
#>  6 1949 Jun   135
#>  7 1949 Jul   148
#>  8 1949 Aug   148
#>  9 1949 Sep   136
#> 10 1949 Oct   119
#> # ... with 134 more rows

as_tsibble(sunspot.year)

#> # A tsibble: 289 x 2 [1YEAR]
#>    index value
#>  * <dbl> <dbl>
#>  1  1700     5
#>  2  1701    11
#>  3  1702    16
#>  4  1703    23
#>  5  1704    36
#>  6  1705    58
#>  7  1706    29
#>  8  1707    20
#>  9  1708    10
#> 10  1709     8
#> # ... with 279 more rows

as_tsibble(sunspot.month)

#> # A tsibble: 3,177 x 2 [1MONTH]
#>       index value
#>  *    <mth> <dbl>
#>  1 1749 Jan  58.0
#>  2 1749 Feb  62.6
#>  3 1749 Mar  70.0
#>  4 1749 Apr  55.7
#>  5 1749 May  85.0
#>  6 1749 Jun  83.5
#>  7 1749 Jul  94.8
#>  8 1749 Aug  66.3
#>  9 1749 Sep  75.9
#> 10 1749 Oct  75.5
#> # ... with 3,167 more rows

as_tsibble(austres)

#> # A tsibble: 89 x 2 [1QUARTER]
#>      index   value
#>  *   <qtr>   <dbl>
#>  1 1971 Q2 13067.3
#>  2 1971 Q3 13130.5
#>  3 1971 Q4 13198.4
#>  4 1972 Q1 13254.2
#>  5 1972 Q2 13303.7
#>  6 1972 Q3 13353.9
#>  7 1972 Q4 13409.3
#>  8 1973 Q1 13459.2
#>  9 1973 Q2 13504.5
#> 10 1973 Q3 13552.6
#> # ... with 79 more rows

# coerce mts to tsibble
z <- ts(matrix(rnorm(300), 100, 3), start = c(1961, 1), frequency = 12)
as_tsibble(z)

#> # A tsibble: 300 x 3 [1MONTH]
#> # Keys:      key
#>       index      key       value
#>  *    <mth>    <chr>       <dbl>
#>  1 1961 Jan Series 1  0.07003485
#>  2 1961 Feb Series 1 -0.63912332
#>  3 1961 Mar Series 1 -0.04996490
#>  4 1961 Apr Series 1 -0.25148344
#>  5 1961 May Series 1  0.44479712
#>  6 1961 Jun Series 1  2.75541758
#>  7 1961 Jul Series 1  0.04653138
#>  8 1961 Aug Series 1  0.57770907
#>  9 1961 Sep Series 1  0.11819487
#> 10 1961 Oct Series 1 -1.91172049
#> # ... with 290 more rows

as_tsibble(z, gather = FALSE)

#> # A tsibble: 100 x 4 [1MONTH]
#>       index  `Series 1` `Series 2`  `Series 3`
#>  *    <mth>       <dbl>      <dbl>       <dbl>
#>  1 1961 Jan  0.07003485 -0.1626763 -0.90281495
#>  2 1961 Feb -0.63912332 -0.8273102  1.31763370
#>  3 1961 Mar -0.04996490  1.8765056  1.10018974
#>  4 1961 Apr -0.25148344  0.7664402  1.20376784
#>  5 1961 May  0.44479712  0.9799567 -1.43127078
#>  6 1961 Jun  2.75541758  1.3217810  1.38291086
#>  7 1961 Jul  0.04653138 -1.1197108  0.00312594
#>  8 1961 Aug  0.57770907  0.5145998 -0.07788682
#>  9 1961 Sep  0.11819487 -1.5090998  0.44142823
#> 10 1961 Oct -1.91172049  1.5327415  0.12892290
#> # ... with 90 more rows

# coerce hts from the "hts" package to tsibble
if (!requireNamespace("hts", quietly = TRUE)) {
  stop("Please install the hts package to run these following examples.")
}
as_tsibble(hts::htseg1)

#> # A tsibble: 50 x 4 [1YEAR]
#> # Keys:      `Level 2` | `Level 1`
#>    index    value `Level 2` `Level 1`
#>  * <dbl>    <dbl>     <chr>     <chr>
#>  1  1992 7.862669        AA         A
#>  2  1993 8.340355        AA         A
#>  3  1994 8.549132        AA         A
#>  4  1995 8.597699        AA         A
#>  5  1996 8.770196        AA         A
#>  6  1997 8.837793        AA         A
#>  7  1998 8.893267        AA         A
#>  8  1999 8.933993        AA         A
#>  9  2000 9.133653        AA         A
#> 10  2001 9.152636        AA         A
#> # ... with 40 more rows

as_tsibble(hts::htseg2)

#> # A tsibble: 160 x 5 [1YEAR]
#> # Keys:      `Level 3` | `Level 2` | `Level 1`
#>    index     value `Level 3` `Level 2` `Level 1`
#>  * <dbl>     <dbl>     <chr>     <chr>     <chr>
#>  1  1992 -2.423474      A10A       A10         A
#>  2  1993 -2.196065      A10A       A10         A
#>  3  1994 -2.187683      A10A       A10         A
#>  4  1995 -1.932308      A10A       A10         A
#>  5  1996 -1.637783      A10A       A10         A
#>  6  1997 -1.627386      A10A       A10         A
#>  7  1998 -1.614443      A10A       A10         A
#>  8  1999 -1.590909      A10A       A10         A
#>  9  2000 -1.532260      A10A       A10         A
#> 10  2001 -1.390682      A10A       A10         A
#> # ... with 150 more rows