Tidyverse methods for tsibble

arrange(): if not arranging key and index in past-to-future order, a warning is likely to be issued.

slice(): if row numbers are not in ascending order, a warning is likely to be issued.

select(): keeps the variables you mention as well as the index.

transmute(): keeps the variable you operate on, as well as the index and key.

summarise() will not collapse on the index variable.

Column-wise verbs, including select(), transmute(), summarise(), mutate() & transmute(), keep the time context hanging around. That is, the index variable cannot be dropped for a tsibble. If any key variable is changed, it will validate whether it's a tsibble internally. Use as_tibble() to leave off the time context.

unnest() requires argument key = id() to get back to a tsibble.

# S3 method for tbl_ts
arrange(.data, ...)

# S3 method for grouped_ts
arrange(.data, ..., .by_group = FALSE)

# S3 method for tbl_ts
filter(.data, ...)

# S3 method for tbl_ts
slice(.data, ...)

# S3 method for tbl_ts
select(.data, ..., .drop = FALSE)

# S3 method for tbl_ts
rename(.data, ...)

# S3 method for tbl_ts
mutate(.data, ..., .drop = FALSE)

# S3 method for tbl_ts
transmute(.data, ..., .drop = FALSE)

# S3 method for tbl_ts
summarise(.data, ..., .drop = FALSE)

# S3 method for tbl_ts
summarize(.data, ..., .drop = FALSE)

# S3 method for tbl_ts
group_by(.data, ..., add = FALSE)

# S3 method for grouped_ts
ungroup(x, ...)

# S3 method for tbl_ts
left_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for tbl_ts
right_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for tbl_ts
inner_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for tbl_ts
full_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for tbl_ts
semi_join(x, y, by = NULL, copy = FALSE, ...)

# S3 method for tbl_ts
anti_join(x, y, by = NULL, copy = FALSE, ...)

# S3 method for tbl_ts
gather(data, key = "key", value = "value", ...,
  na.rm = FALSE, convert = FALSE, factor_key = FALSE)

# S3 method for tbl_ts
spread(data, key, value, fill = NA, convert = FALSE,
  drop = TRUE, sep = NULL)

# S3 method for tbl_ts
nest(data, ..., .key = "data")

# S3 method for lst_ts
unnest(data, ..., key = id(), .drop = NA,
  .id = NULL, .sep = NULL, .preserve = NULL)

# S3 method for tbl_ts
unnest(data, ..., key = id(), .drop = NA,
  .id = NULL, .sep = NULL, .preserve = NULL)

# S3 method for grouped_ts
fill(data, ..., .direction = c("down", "up"))

Arguments

.data
A tbl_ts.

...
same arguments accepted as its dplyr generic.

.by_group
If TRUE, will sort first by grouping variable. Applies to grouped data frames only.

.drop
Deprecated, please use as_tibble() for .drop = TRUE instead. FALSE returns a tsibble object as the input. TRUE drops a tsibble and returns a tibble.

add
When add = FALSE, the default, group_by() will override existing groups. To add to the existing groups, use add = TRUE.

x
A tbl()

y
tbls to join

by
a character vector of variables to join by. If NULL, the default, *_join() will do a natural join, using all variables with common names across the two tables. A message lists the variables so that you can check they're right (to suppress the message, simply explicitly list the variables that you want to join).

To join by different variables on x and y use a named vector. For example, by = c("a" = "b") will match x.a to y.b.

copy
If x and y are not from the same data source, and copy is TRUE, then y will be copied into the same src as x. This allows you to join tables across srcs, but it is a potentially expensive operation so you must opt into it.

suffix
If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2.

data
A data frame.

key
Unquoted variables to create the key (via id) after unnesting.

value
Names of new key and value columns, as strings or symbols.

This argument is passed by expression and supports quasiquotation (you can unquote strings and symbols). The name is captured from the expression with rlang::ensym() (note that this kind of interface where symbols do not represent actual objects is now discouraged in the tidyverse; we support it here for backward compatibility).

na.rm
If TRUE, will remove rows from output where the value column in NA.

convert
If TRUE will automatically run type.convert() on the key column. This is useful if the column types are actually numeric, integer, or logical.

factor_key
If FALSE, the default, the key values will be stored as a character vector. If TRUE, will be stored as a factor, which preserves the original ordering of the columns.

fill
If set, missing values will be replaced with this value. Note that there are two types of missingness in the input: explicit missing values (i.e. NA), and implicit missings, rows that simply aren't present. Both types of missing value will be replaced by fill.

drop
If FALSE, will keep factor levels that don't appear in the data, filling in missing combinations with fill.

sep
If NULL, the column names will be taken from the values of key variable. If non-NULL, the column names will be given by "<key_name><sep><key_value>".

.key
The name of the new column, as a string or symbol.

This argument is passed by expression and supports quasiquotation (you can unquote strings and symbols). The name is captured from the expression with rlang::ensym() (note that this kind of interface where symbols do not represent actual objects is now discouraged in the tidyverse; we support it here for backward compatibility).

.id
Data frame identifier - if supplied, will create a new column with name .id, giving a unique identifier. This is most useful if the list column is named.

.sep
If non-NULL, the names of unnested data frame columns will combine the name of the original list-col with the names from nested data frame, separated by .sep.

.preserve
Optionally, list-columns to preserve in the output. These will be duplicated in the same way as atomic vectors. This has dplyr::select semantics so you can preserve multiple variables with .preserve = c(x, y) or .preserve = starts_with("list").

.direction
Direction in which to fill missing values. Currently either "down" (the default) or "up".

Examples

# Sum over sensors ----
pedestrian %>%
  summarise(Total = sum(Count))

#> # A tsibble: 17,542 x 2 [1h] <Australia/Melbourne>
#>    Date_Time           Total
#>    <dttm>              <int>
#>  1 2015-01-01 00:00:00  2866
#>  2 2015-01-01 01:00:00  1535
#>  3 2015-01-01 02:00:00   994
#>  4 2015-01-01 03:00:00   569
#>  5 2015-01-01 04:00:00   311
#>  6 2015-01-01 05:00:00   159
#>  7 2015-01-01 06:00:00   129
#>  8 2015-01-01 07:00:00   146
#>  9 2015-01-01 08:00:00   258
#> 10 2015-01-01 09:00:00   419
#> # … with 17,532 more rows

# Back to tibble
pedestrian %>%
  as_tibble() %>%
  summarise(Total = sum(Count))

#> # A tibble: 1 x 1
#>      Total
#>      <int>
#> 1 45483871

# example from tidyr
stocks <- tsibble(
  time = as.Date('2009-01-01') + 0:9,
  X = rnorm(10, 0, 1),
  Y = rnorm(10, 0, 2),
  Z = rnorm(10, 0, 4)
)

#> Column `time` is the index.

stocks %>% gather(stock, price, -time)

#> # A tsibble: 30 x 3 [1D]
#> # Key:       stock [3]
#>    time       stock  price
#>    <date>     <chr>  <dbl>
#>  1 2009-01-01 X      0.168
#>  2 2009-01-02 X      1.18 
#>  3 2009-01-03 X      1.53 
#>  4 2009-01-04 X     -0.565
#>  5 2009-01-05 X      0.296
#>  6 2009-01-06 X     -0.249
#>  7 2009-01-07 X     -0.207
#>  8 2009-01-08 X      0.822
#>  9 2009-01-09 X      0.704
#> 10 2009-01-10 X     -0.158
#> # … with 20 more rows

# example from tidyr
stocks <- tsibble(
  time = as.Date('2009-01-01') + 0:9,
  X = rnorm(10, 0, 1),
  Y = rnorm(10, 0, 2),
  Z = rnorm(10, 0, 4)
)

#> Column `time` is the index.

stocksm <- stocks %>% gather(stock, price, -time)
stocksm %>% spread(stock, price)

#> # A tsibble: 10 x 4 [1D]
#>    time            X      Y      Z
#>    <date>      <dbl>  <dbl>  <dbl>
#>  1 2009-01-01  0.642  1.22   0.340
#>  2 2009-01-02 -0.417 -2.00   2.33 
#>  3 2009-01-03 -1.22   2.39   1.80 
#>  4 2009-01-04  2.13  -3.55   1.99 
#>  5 2009-01-05 -1.09   1.57   3.12 
#>  6 2009-01-06  0.116 -0.875  6.03 
#>  7 2009-01-07  1.26  -0.703 -7.30 
#>  8 2009-01-08  1.04   2.61   3.15 
#>  9 2009-01-09 -0.969 -2.17  -3.89 
#> 10 2009-01-10  0.391 -2.45   3.57 

nested_stock <- stocksm %>%
  nest(-stock)
stocksm %>%
  group_by(stock) %>%
  nest()

#> # A tibble: 3 x 2
#>   stock data              
#>   <chr> <list>            
#> 1 X     <tsibble [10 × 2]>
#> 2 Y     <tsibble [10 × 2]>
#> 3 Z     <tsibble [10 × 2]>

nested_stock %>%
  unnest(key = id(stock))

#> # A tsibble: 30 x 3 [1D]
#> # Key:       stock [3]
#>    stock time        price
#>    <chr> <date>      <dbl>
#>  1 X     2009-01-01  0.642
#>  2 X     2009-01-02 -0.417
#>  3 X     2009-01-03 -1.22 
#>  4 X     2009-01-04  2.13 
#>  5 X     2009-01-05 -1.09 
#>  6 X     2009-01-06  0.116
#>  7 X     2009-01-07  1.26 
#>  8 X     2009-01-08  1.04 
#>  9 X     2009-01-09 -0.969
#> 10 X     2009-01-10  0.391
#> # … with 20 more rows

stock_qtl <- stocksm %>%
  group_by(stock) %>%
  index_by(day3 = lubridate::floor_date(time, unit = "3 day")) %>%
  summarise(
    value = list(quantile(price)),
    qtl = list(c("0%", "25%", "50%", "75%", "100%"))
  )
unnest(stock_qtl, key = id(qtl))

#> # A tsibble: 60 x 4 [3D]
#> # Key:       stock, qtl [15]
#>    stock day3        value qtl  
#>    <chr> <date>      <dbl> <chr>
#>  1 X     2009-01-01 -1.22  0%   
#>  2 X     2009-01-01 -0.817 25%  
#>  3 X     2009-01-01 -0.417 50%  
#>  4 X     2009-01-01  0.112 75%  
#>  5 X     2009-01-01  0.642 100% 
#>  6 X     2009-01-04 -1.09  0%   
#>  7 X     2009-01-04 -0.486 25%  
#>  8 X     2009-01-04  0.116 50%  
#>  9 X     2009-01-04  1.12  75%  
#> 10 X     2009-01-04  2.13  100% 
#> # … with 50 more rows