Tidyverse methods for tsibble

arrange(): if not arranging key and index in past-to-future order, a warning is likely to be issued.

slice(): if row numbers are not in ascending order, a warning is likely to be issued.

select(): keeps the variables you mention as well as the index.

transmute(): keeps the variable you operate on, as well as the index and key.

summarise() will not collapse on the index variable.

Column-wise verbs, including select(), transmute(), summarise(), mutate() & transmute(), keep the time context hanging around. That is, the index variable cannot be dropped for a tsibble. If any key variable is changed, it will validate whether it's a tsibble internally. Use as_tibble() to leave off the time context.

# S3 method for tbl_ts
arrange(.data, ...)

# S3 method for grouped_ts
arrange(.data, ..., .by_group = FALSE)

# S3 method for tbl_ts
filter(.data, ...)

# S3 method for tbl_ts
slice(.data, ...)

# S3 method for tbl_ts
select(.data, ..., .drop = FALSE)

# S3 method for tbl_ts
rename(.data, ...)

# S3 method for tbl_ts
mutate(.data, ..., .drop = FALSE)

# S3 method for tbl_ts
transmute(.data, ..., .drop = FALSE)

# S3 method for tbl_ts
summarise(.data, ..., .drop = FALSE)

# S3 method for tbl_ts
summarize(.data, ..., .drop = FALSE)

# S3 method for tbl_ts
group_by(.data, ..., add = FALSE)

# S3 method for grouped_ts
ungroup(x, ...)

# S3 method for tbl_ts
left_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for tbl_ts
right_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for tbl_ts
inner_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for tbl_ts
full_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for tbl_ts
semi_join(x, y, by = NULL, copy = FALSE, ...)

# S3 method for tbl_ts
anti_join(x, y, by = NULL, copy = FALSE, ...)

# S3 method for tbl_ts
gather(data, key = "key", value = "value", ...,
  na.rm = FALSE, convert = FALSE, factor_key = FALSE)

# S3 method for tbl_ts
spread(data, key, value, fill = NA, convert = FALSE,
  drop = TRUE, sep = NULL)

# S3 method for tbl_ts
nest(data, ..., .key = "data")

# S3 method for lst_ts
unnest(data, ..., key = id(), .drop = NA,
  .id = NULL, .sep = NULL, .preserve = NULL)

Arguments

.data
A tbl_ts.

...
same arguments accepted as its dplyr generic.

.by_group
If TRUE, will sort first by grouping variable. Applies to grouped data frames only.

.drop
Deprecated, please use as_tibble() for .drop = TRUE instead. FALSE returns a tsibble object as the input. TRUE drops a tsibble and returns a tibble.

add
When add = FALSE, the default, group_by() will override existing groups. To add to the existing groups, use add = TRUE.

x
A tbl()

y
tbls to join

by
a character vector of variables to join by. If NULL, the default, *_join() will do a natural join, using all variables with common names across the two tables. A message lists the variables so that you can check they're right (to suppress the message, simply explicitly list the variables that you want to join).

To join by different variables on x and y use a named vector. For example, by = c("a" = "b") will match x.a to y.b.

copy
If x and y are not from the same data source, and copy is TRUE, then y will be copied into the same src as x. This allows you to join tables across srcs, but it is a potentially expensive operation so you must opt into it.

suffix
If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2.

data
A data frame.

key
Unquoted variables to create the key (via id) after unnesting.

value
Names of new key and value columns, as strings or symbols.

This argument is passed by expression and supports quasiquotation (you can unquote strings and symbols). The name is captured from the expression with rlang::ensym() (note that this kind of interface where symbols do not represent actual objects is now discouraged in the tidyverse; we support it here for backward compatibility).

na.rm
If TRUE, will remove rows from output where the value column in NA.

convert
If TRUE will automatically run type.convert() on the key column. This is useful if the column types are actually numeric, integer, or logical.

factor_key
If FALSE, the default, the key values will be stored as a character vector. If TRUE, will be stored as a factor, which preserves the original ordering of the columns.

fill
If set, missing values will be replaced with this value. Note that there are two types of missingness in the input: explicit missing values (i.e. NA), and implicit missings, rows that simply aren't present. Both types of missing value will be replaced by fill.

drop
If FALSE, will keep factor levels that don't appear in the data, filling in missing combinations with fill.

sep
If NULL, the column names will be taken from the values of key variable. If non-NULL, the column names will be given by "<key_name><sep><key_value>".

.key
The name of the new column, as a string or symbol.

This argument is passed by expression and supports quasiquotation (you can unquote strings and symbols). The name is captured from the expression with rlang::ensym() (note that this kind of interface where symbols do not represent actual objects is now discouraged in the tidyverse; we support it here for backward compatibility).

.id
Data frame identifier - if supplied, will create a new column with name .id, giving a unique identifier. This is most useful if the list column is named.

.sep
If non-NULL, the names of unnested data frame columns will combine the name of the original list-col with the names from nested data frame, separated by .sep.

.preserve
Optionally, list-columns to preserve in the output. These will be duplicated in the same way as atomic vectors. This has dplyr::select semantics so you can preserve multiple variables with .preserve = c(x, y) or .preserve = starts_with("list").

Examples

# Sum over sensors ----
pedestrian %>%
  summarise(Total = sum(Count))

#> # A tsibble: 17,542 x 2 [1h]
#>    Date_Time           Total
#>    <dttm>              <int>
#>  1 2015-01-01 00:00:00  2866
#>  2 2015-01-01 01:00:00  1535
#>  3 2015-01-01 02:00:00   994
#>  4 2015-01-01 03:00:00   569
#>  5 2015-01-01 04:00:00   311
#>  6 2015-01-01 05:00:00   159
#>  7 2015-01-01 06:00:00   129
#>  8 2015-01-01 07:00:00   146
#>  9 2015-01-01 08:00:00   258
#> 10 2015-01-01 09:00:00   419
#> # ... with 17,532 more rows

# Sum over sensors by days ----
pedestrian %>%
  index_by(Date) %>%
  summarise(Total = sum(Count))

#> # A tsibble: 731 x 2 [1D]
#>    Date       Total
#>    <date>     <int>
#>  1 2015-01-01 18079
#>  2 2015-01-02 19306
#>  3 2015-01-03 12708
#>  4 2015-01-04 20665
#>  5 2015-01-05 29724
#>  6 2015-01-06 31024
#>  7 2015-01-07 28842
#>  8 2015-01-08 28664
#>  9 2015-01-09 31269
#> 10 2015-01-10 30904
#> # ... with 721 more rows

# Back to tibble
pedestrian %>%
  as_tibble() %>%
  summarise(Total = sum(Count))

#> # A tibble: 1 x 1
#>      Total
#>      <int>
#> 1 45483871

tourism %>%
  group_by(Region, State) %>%
  summarise(geo_trips = sum(Trips))

#> # A tsibble: 5,852 x 4 [1Q]
#> # Key:       Region, State [77]
#> # Groups:    Region [77]
#>    Region   State           Quarter geo_trips
#>    <chr>    <chr>             <qtr>     <dbl>
#>  1 Adelaide South Australia 1998 Q1      659.
#>  2 Adelaide South Australia 1998 Q2      450.
#>  3 Adelaide South Australia 1998 Q3      593.
#>  4 Adelaide South Australia 1998 Q4      524.
#>  5 Adelaide South Australia 1999 Q1      548.
#>  6 Adelaide South Australia 1999 Q2      569.
#>  7 Adelaide South Australia 1999 Q3      538.
#>  8 Adelaide South Australia 1999 Q4      562.
#>  9 Adelaide South Australia 2000 Q1      646.
#> 10 Adelaide South Australia 2000 Q2      563.
#> # ... with 5,842 more rows

# example from tidyr
stocks <- tsibble(
  time = as.Date('2009-01-01') + 0:9,
  X = rnorm(10, 0, 1),
  Y = rnorm(10, 0, 2),
  Z = rnorm(10, 0, 4)
)

#> Column `time` is the index.

stocks %>% gather(stock, price, -time)

#> # A tsibble: 30 x 3 [1D]
#> # Key:       stock [3]
#>    time       stock  price
#>    <date>     <chr>  <dbl>
#>  1 2009-01-01 X      0.637
#>  2 2009-01-02 X      0.514
#>  3 2009-01-03 X     -1.75 
#>  4 2009-01-04 X      0.894
#>  5 2009-01-05 X      0.223
#>  6 2009-01-06 X      0.581
#>  7 2009-01-07 X     -0.178
#>  8 2009-01-08 X      0.741
#>  9 2009-01-09 X     -0.997
#> 10 2009-01-10 X     -2.94 
#> # ... with 20 more rows

# example from tidyr
stocks <- tsibble(
  time = as.Date('2009-01-01') + 0:9,
  X = rnorm(10, 0, 1),
  Y = rnorm(10, 0, 2),
  Z = rnorm(10, 0, 4)
)

#> Column `time` is the index.

stocksm <- stocks %>% gather(stock, price, -time)
stocksm %>% spread(stock, price)

#> # A tsibble: 10 x 4 [1D]
#>    time             X        Y      Z
#>    <date>       <dbl>    <dbl>  <dbl>
#>  1 2009-01-01 -0.323  -4.17    -1.10 
#>  2 2009-01-02 -1.29    0.00714 -2.37 
#>  3 2009-01-03 -1.03   -0.712   -0.489
#>  4 2009-01-04  1.51    2.29     4.72 
#>  5 2009-01-05  0.347  -0.442    2.56 
#>  6 2009-01-06  1.78    2.04    -2.52 
#>  7 2009-01-07  0.387  -0.527   -3.23 
#>  8 2009-01-08 -0.919   3.32    -3.44 
#>  9 2009-01-09 -1.58   -1.55    -8.68 
#> 10 2009-01-10 -0.0841 -1.85    -5.50 

pedestrian %>%
  nest(-Sensor)

#> # A tibble: 4 x 2
#>   Sensor                        data                  
#>   <chr>                         <list>                
#> 1 Birrarung Marr                <tsibble [14,566 × 4]>
#> 2 Bourke Street Mall (North)    <tsibble [16,414 × 4]>
#> 3 QV Market-Elizabeth St (West) <tsibble [17,518 × 4]>
#> 4 Southern Cross Station        <tsibble [17,539 × 4]>

pedestrian %>%
  group_by(Sensor) %>%
  nest()

#> # A tibble: 4 x 2
#>   Sensor                        data                  
#>   <chr>                         <list>                
#> 1 Birrarung Marr                <tsibble [14,566 × 4]>
#> 2 Bourke Street Mall (North)    <tsibble [16,414 × 4]>
#> 3 QV Market-Elizabeth St (West) <tsibble [17,518 × 4]>
#> 4 Southern Cross Station        <tsibble [17,539 × 4]>

nested_ped <- pedestrian %>%
  nest(-Sensor)
nested_ped %>%
  unnest(key = id(Sensor))

#> # A tsibble: 66,037 x 5 [1h]
#> # Key:       Sensor [4]
#>    Sensor         Date_Time           Date        Time Count
#>    <chr>          <dttm>              <date>     <int> <int>
#>  1 Birrarung Marr 2015-01-01 00:00:00 2015-01-01     0  1630
#>  2 Birrarung Marr 2015-01-01 01:00:00 2015-01-01     1   826
#>  3 Birrarung Marr 2015-01-01 02:00:00 2015-01-01     2   567
#>  4 Birrarung Marr 2015-01-01 03:00:00 2015-01-01     3   264
#>  5 Birrarung Marr 2015-01-01 04:00:00 2015-01-01     4   139
#>  6 Birrarung Marr 2015-01-01 05:00:00 2015-01-01     5    77
#>  7 Birrarung Marr 2015-01-01 06:00:00 2015-01-01     6    44
#>  8 Birrarung Marr 2015-01-01 07:00:00 2015-01-01     7    56
#>  9 Birrarung Marr 2015-01-01 08:00:00 2015-01-01     8   113
#> 10 Birrarung Marr 2015-01-01 09:00:00 2015-01-01     9   166
#> # ... with 66,027 more rows

nested_tourism <- tourism %>%
  nest(-Region, -State)
nested_tourism %>%
  unnest(key = id(Region, State))

#> # A tsibble: 23,408 x 5 [1Q]
#> # Key:       Purpose, Region, State [308]
#>    Region   State           Quarter Purpose  Trips
#>    <chr>    <chr>             <qtr> <chr>    <dbl>
#>  1 Adelaide South Australia 1998 Q1 Business  135.
#>  2 Adelaide South Australia 1998 Q2 Business  110.
#>  3 Adelaide South Australia 1998 Q3 Business  166.
#>  4 Adelaide South Australia 1998 Q4 Business  127.
#>  5 Adelaide South Australia 1999 Q1 Business  137.
#>  6 Adelaide South Australia 1999 Q2 Business  200.
#>  7 Adelaide South Australia 1999 Q3 Business  169.
#>  8 Adelaide South Australia 1999 Q4 Business  134.
#>  9 Adelaide South Australia 2000 Q1 Business  154.
#> 10 Adelaide South Australia 2000 Q2 Business  169.
#> # ... with 23,398 more rows