Create a tsibble object

tsibble(..., key = NULL, index, regular = TRUE, .drop = TRUE)

Arguments

...
A set of name-value pairs. The names of "key" and "index" should be avoided as they are used as the arguments.

key
Unquoted variable(s) that uniquely determine time indices. NULL for empty key, and works with tidy selector (e.g. dplyr::starts_with()).

index
A bare (or unquoted) variable to specify the time index variable.

regular
Regular time interval (TRUE) or irregular (FALSE). The interval is determined by the greatest common divisor of index column, if TRUE.

.drop
If TRUE, empty key groups are dropped.

Value

A tsibble object.

Details

A tsibble is sorted by its key first and index.

Index

An extensive range of indices are supported by tsibble: native time classes in R (such as Date, POSIXct, and difftime) and tsibble's new additions (such as yearweek, yearmonth, and yearquarter). Some commonly-used classes have built-in support too, including ordered, hms::hms, zoo::yearmon, zoo::yearqtr, and nanotime.

For a tbl_ts of regular interval, a choice of index representation has to be made. For example, a monthly data should correspond to time index created by yearmonth or zoo::yearmon, instead of Date or POSIXct. Because months in a year ensures the regularity, 12 months every year. However, if using Date, a month containing days ranges from 28 to 31 days, which results in irregular time space. This is also applicable to year-week and year-quarter.

Since the tibble that underlies the tsibble only accepts a 1d atomic vector or a list, the tsibble doesn't accept types of POSIXlt and timeDate.

Tsibble supports arbitrary index classes, as long as they can be ordered from past to future. To support a custom class, one needs to define index_valid() for the class and calculate the interval through interval_pull().

Key

Key variable(s) together with the index uniquely identifies each record:

Empty: an implicit variable. NULL resulting in a univariate time series.

A single variable: For example, data(pedestrian) use the bare Sensor as the key.

Multiple variables: For example, Declare key = c(Region, State, Purpose) for data(tourism). Key can be created in conjunction with tidy selectors like starts_with().

Interval

The interval function returns the interval associated with the tsibble.

Regular: the value and its time unit including "nanosecond", "microsecond", "millisecond", "second", "minute", "hour", "day", "week", "month", "quarter", "year". An unrecognisable time interval is labelled as "unit".

Irregular: as_tsibble(regular = FALSE) gives the irregular tsibble. It is marked with !.

Unknown: if there is only one entry for each key variable, the interval cannot be determined (?).

An interval is obtained based on the corresponding index representation:

integerish numerics between 1582 and 2499: "year" (Y). Note the year of 1582 saw the beginning of the Gregorian Calendar switch.

yearquarter/yearqtr: "quarter" (Q)

yearmonth/yearmon: "month" (M)

yearweek: "week" (W)

Date: "day" (D)

difftime: "quarter" (Q), "month" (M), "week" (W), "day" (D), "hour" (h), "minute" (m), "second" (s)

POSIXct/hms: "hour" (h), "minute" (m), "second" (s), "millisecond" (us), "microsecond" (ms)

nanotime: "nanosecond" (ns)

other numerics &ordered (ordered factor): "unit"

See also

Examples

# create a tsibble w/o a key
tsibble(
  date = as.Date("2017-01-01") + 0:9,
  value = rnorm(10)
)

#> Using `date` as index variable.

#> # A tsibble: 10 x 2 [1D]
#>    date         value
#>    <date>       <dbl>
#>  1 2017-01-01  2.52  
#>  2 2017-01-02 -0.420 
#>  3 2017-01-03  1.34  
#>  4 2017-01-04 -0.223 
#>  5 2017-01-05  0.593 
#>  6 2017-01-06 -0.0653
#>  7 2017-01-07 -0.701 
#>  8 2017-01-08 -0.529 
#>  9 2017-01-09  0.170 
#> 10 2017-01-10  0.114 

# create a tsibble with one key
tsibble(
  qtr = rep(yearquarter("201001") + 0:9, 3),
  group = rep(c("x", "y", "z"), each = 10),
  value = rnorm(30),
  key = group
)

#> Using `qtr` as index variable.

#> # A tsibble: 30 x 3 [1Q]
#> # Key:       group [3]
#>        qtr group   value
#>      <qtr> <chr>   <dbl>
#>  1 2010 Q1 x     -0.796 
#>  2 2010 Q2 x     -0.236 
#>  3 2010 Q3 x      1.19  
#>  4 2010 Q4 x      0.188 
#>  5 2011 Q1 x      1.04  
#>  6 2011 Q2 x     -0.450 
#>  7 2011 Q3 x     -0.709 
#>  8 2011 Q4 x      0.0304
#>  9 2012 Q1 x     -0.450 
#> 10 2012 Q2 x      0.671 
#> # … with 20 more rows